requirements.txt
absl-py==2.1.0
arcface==0.0.8
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepface==0.0.92
distlib==0.3.8
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
gast==0.4.0
gdown==5.2.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.65.1
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==2.13.1
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
mtcnn==0.1.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
platformdirs==4.2.2
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9
six==1.16.0
soupsieve==2.5
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.4
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
